+++ src/main/java/org/calpoly/gehackathon/domain/Tree.java
@@ -1,6 +1,4 @@
-package org.cloudfoundry.samples.music.domain;
-
-import org.hibernate.annotations.GenericGenerator;
+package org.calpoly.gehackathon.domain;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -9,16 +7,9 @@ import javax.persistence.Id;
 
 @Entity
 public class Tree {
-
-
-    public String getId() {
-        return id;
-    }
-
     @Id
     @Column(length=40)
-    @GeneratedValue(generator="randomId")
-    @GenericGenerator(name="randomId", strategy="org.cloudfoundry.samples.music.domain.RandomIdGenerator")
+    @GeneratedValue
     private String id;
 
     public String name;
@@ -26,14 +17,17 @@ public class Tree {
     public Double xCoord;
     public Double yCoord;
 
-
-
     public Tree() {
     }
+
     public Tree(String name, String type, Double xCoord, Double yCoord) {
         this.name = name;
         this.type = type;
         this.xCoord = xCoord;
         this.yCoord = yCoord;
+    }
+
+    public String getId() {
+        return id;
     }
 }
