+++ src/main/java/org/calpoly/gehackathon/repositories/jpa/JpaPedestrianRepository.java
@@ -1,7 +1,7 @@
-package org.cloudfoundry.samples.music.repositories.jpa;
+package org.calpoly.gehackathon.repositories.jpa;
 
-import org.cloudfoundry.samples.music.config.SpringApplicationContextInitializer;
-import org.cloudfoundry.samples.music.domain.Tree;
+import org.calpoly.gehackathon.config.SpringApplicationContextInitializer;
+import org.calpoly.gehackathon.domain.Pedestrian;
 import org.springframework.context.annotation.Profile;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -10,5 +10,5 @@ import org.springframework.stereotype.Repository;
 @Profile({SpringApplicationContextInitializer.LOCAL_PROFILE,
           SpringApplicationContextInitializer.IN_MEMORY_PROFILE,
           SpringApplicationContextInitializer.CLOUD_PROFILE})
-public interface JpaTreeRepository extends JpaRepository<Tree, String> {
+public interface JpaPedestrianRepository extends JpaRepository<Pedestrian, Integer> {
 }
