+++ src/main/java/org/calpoly/gehackathon/web/PedestrianController.java
@@ -1,7 +1,7 @@
-package org.cloudfoundry.samples.music.web;
+package org.calpoly.gehackathon.web;
 
-import org.cloudfoundry.samples.music.domain.Pedestrian;
-import org.cloudfoundry.samples.music.repositories.jpa.JpaPedestrianRepository;
+import org.calpoly.gehackathon.domain.Pedestrian;
+import org.calpoly.gehackathon.repositories.jpa.JpaPedestrianRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
